+++ packages/atomic/src/components/atomic-result-template/atomic-result-template.tsx
@@ -43,8 +43,10 @@ export class AtomicResultTemplate {
   @MapProp() public mustNotMatch: Record<string, string[]> = {};
 
   constructor() {
+    // TODO: Remove second half of the condition in v1
     const isParentResultList =
-      this.host.parentElement?.nodeName === 'ATOMIC-RESULT-LIST';
+      this.host.parentElement?.nodeName === 'ATOMIC-RESULT-LIST' ||
+      this.host.parentElement?.nodeName === 'ATOMIC-RESULT-LIST-V1';
 
     if (!isParentResultList) {
       this.error = new Error(