+++ packages/atomic/src/components/facets-v1/facet-header/facet-header.tsx
@@ -10,7 +10,7 @@ export const FacetHeader: FunctionalComponent<{
   numberOfSelectedValues: number;
   isCollapsed: boolean;
   onToggleCollapse(): void;
-  onClearFilters(): void;
+  onClearFilters?(): void;
 }> = (props) => {
   const label = props.i18n.t(props.label);
   const expandFacet = props.i18n.t('expand-facet', {label});
@@ -26,24 +26,24 @@ export const FacetHeader: FunctionalComponent<{
   return [
     <button
       part="label-button"
-      class="flex justify-between w-full py-1 text-on-background text-lg hover:text-primary"
+      class="flex justify-between w-full py-1 px-2 text-on-background text-lg hover:text-primary focus:text-primary"
       title={props.isCollapsed ? expandFacet : collapseFacet}
       onClick={() => props.onToggleCollapse()}
       aria-expanded={(!props.isCollapsed).toString()}
     >
-      <span class="ellipsed">{label}</span>
+      <span class="truncate">{label}</span>
       <span
         part="label-button-icon"
         class="fill-current w-3 h-2 self-center flex-shrink-0 ml-4"
         innerHTML={props.isCollapsed ? ArrowTopIcon : ArrowBottomIcon}
       ></span>
     </button>,
-    props.numberOfSelectedValues > 0 && (
+    props.onClearFilters && props.numberOfSelectedValues > 0 && (
       <button
         part="clear-button"
-        class="flex items-baseline w-full p-1 text-sm link"
+        class="flex items-baseline p-2 text-sm text-primary rounded hover:text-primary focus:text-primary hover:bg-neutral-light focus:bg-neutral-light"
         title={clearFiltersForFacet}
-        onClick={() => props.onClearFilters()}
+        onClick={() => props.onClearFilters!()}
       >
         <span
           part="clear-button-icon"