+++ packages/atomic/src/components/facets-v1/facet-value-checkbox/facet-value-checkbox.tsx
@@ -14,30 +14,23 @@ export const FacetValueCheckbox: FunctionalComponent<FacetValueProps> = (
   });
 
   return (
-    <li key={props.displayValue} class="flex items-center">
+    <li key={props.displayValue} class="relative flex items-center">
       <button
         id={id}
         role="checkbox"
         part="value-checkbox"
         onClick={() => props.onClick()}
         aria-checked={props.isSelected.toString()}
-        class={`value-checkbox flex justify-center rounded focus:outline-none focus:border-primary-light ${
+        class={`value-checkbox ${
           props.isSelected
             ? 'selected bg-primary'
             : 'border border-neutral-dark'
         }`}
         aria-label={ariaLabel}
       ></button>
-      <label
-        htmlFor={id}
-        part="value-checkbox-label"
-        class="w-full flex items-center pl-2 py-2.5 text-on-background cursor-pointer ellipsed"
-      >
+      <label htmlFor={id} part="value-checkbox-label">
         {children}
-        <span
-          part="value-count"
-          class="ml-1.5 text-neutral-dark with-parentheses"
-        >
+        <span part="value-count" class="value-count">
           {count}
         </span>
       </label>