+++ packages/atomic/tailwind.config.js
@@ -38,6 +38,10 @@ module.exports = {
         base: 'var(--atomic-text-base)',
         lg: 'var(--atomic-text-lg)',
       },
+      screens: {
+        'desktop-only': {min: '768px'},
+        'mobile-only': {max: '767px'},
+      },
       gridTemplateColumns: {
         'min-1fr': 'min-content 1fr',
       },
@@ -51,7 +55,8 @@ module.exports = {
   },
   variants: {
     extend: {
-      textColor: ['visited'],
+      textColor: ['visited', 'group-focus'],
+      backgroundColor: ['group-focus'],
     },
   },
   plugins: [],